Move `EventStorage` to `event` package
