+++ server/src/main/java/org/spine3/server/storage/StorageFactory.java
@@ -25,6 +25,7 @@ import org.spine3.server.aggregate.AggregateStorage;
 import org.spine3.server.command.CommandStorage;
 import org.spine3.server.command.CommandStore;
 import org.spine3.server.entity.Entity;
+import org.spine3.server.event.EventStorage;
 import org.spine3.server.event.EventStore;
 import org.spine3.server.stand.Stand;
