+++ server/src/main/java/org/spine3/server/storage/memory/InMemoryEventStorage.java
@@ -26,8 +26,8 @@ import com.google.protobuf.Timestamp;
 import org.spine3.base.Event;
 import org.spine3.base.EventId;
 import org.spine3.protobuf.Timestamps;
+import org.spine3.server.event.EventStorage;
 import org.spine3.server.event.EventStreamQuery;
-import org.spine3.server.storage.EventStorage;
 import org.spine3.server.storage.EventStorageRecord;
 
 import javax.annotation.Nullable;
