+++ server/src/main/java/org/spine3/server/storage/memory/InMemoryStorageFactory.java
@@ -24,7 +24,7 @@ import org.spine3.server.aggregate.Aggregate;
 import org.spine3.server.aggregate.AggregateStorage;
 import org.spine3.server.command.CommandStorage;
 import org.spine3.server.entity.Entity;
-import org.spine3.server.storage.EventStorage;
+import org.spine3.server.event.EventStorage;
 import org.spine3.server.storage.ProjectionStorage;
 import org.spine3.server.storage.RecordStorage;
 import org.spine3.server.storage.StandStorage;
