+++ server/src/test/java/org/spine3/server/event/EventStoreServiceBuilderShould.java
@@ -23,7 +23,6 @@ package org.spine3.server.event;
 import com.google.common.util.concurrent.MoreExecutors;
 import org.junit.Before;
 import org.junit.Test;
-import org.spine3.server.storage.EventStorage;
 import org.spine3.server.storage.memory.InMemoryStorageFactory;
 import org.spine3.test.Tests;
